Test placement for Corporate Identity under Association buttons
This is an initial test for adding the Corporate Identity button to the Members page as requested by STIJL.

I have put this under the Association tag as it is directly related to GEWIS itself. The only reason to put it under Manage would be the formatting, as 5 buttons with one extra is far from ideal.

The current GEWIS logo is from the STIJL website, but has some issues with the smoothness of the edges.