Introduce module specific `LocalisedText`
This change makes the old `LocalisedText` an abstract class in
the `Application` module and two new classes which "extend" this
abstract class are added to the specific modules (`Activity` and
`Company`).

I have chosen to do it like this, as with a Doctrine
`InheritanceType` there would either be a single table for all
`LocalisedText` or an additional table which would cause
additional `JOIN`-statements when querying the database.